requirements.txt
pillow
psycopg2-binary